Bump setuptools to 76.0.0
