+++ stubs/setuptools/METADATA.toml
@@ -1,4 +1,4 @@
-version = "~=75.8.2"
+version = "~=76.0.0"
 upstream_repository = "https://github.com/pypa/setuptools"
 extra_description = """\
 Given that `pkg_resources` is typed since `setuptools >= 71.1`, \
