+++ stubs/setuptools/distutils/compilers/C/base.pyi
@@ -1,0 +1,1 @@
+from setuptools._distutils.compilers.C.base import *
