+++ stubs/setuptools/distutils/compilers/C/errors.pyi
@@ -1,0 +1,1 @@
+from setuptools._distutils.compilers.C.errors import *
