+++ stubs/setuptools/distutils/compilers/C/msvc.pyi
@@ -1,0 +1,1 @@
+from setuptools._distutils.compilers.C.msvc import *
