+++ stubs/setuptools/setuptools/_distutils/compilers/C/errors.pyi
@@ -1,0 +1,7 @@
+class Error(Exception): ...
+class PreprocessError(Error): ...
+class CompileError(Error): ...
+class LibError(Error): ...
+class LinkError(Error): ...
+class UnknownFileError(Error): ...
+class UnknownFileType(Error): ...
