+++ stubs/setuptools/setuptools/_distutils/errors.pyi
@@ -1,3 +1,15 @@
+from .compilers.C.errors import (
+    CompileError as CompileError,
+    Error as _Error,
+    LibError as LibError,
+    LinkError as LinkError,
+    PreprocessError as PreprocessError,
+    UnknownFileType as _UnknownFileType,
+)
+
+CCompilerError = _Error
+UnknownFileError = _UnknownFileType
+
 class DistutilsError(Exception): ...
 class DistutilsModuleError(DistutilsError): ...
 class DistutilsClassError(DistutilsError): ...
@@ -11,9 +23,3 @@ class DistutilsExecError(DistutilsError): ...
 class DistutilsInternalError(DistutilsError): ...
 class DistutilsTemplateError(DistutilsError): ...
 class DistutilsByteCompileError(DistutilsError): ...
-class CCompilerError(Exception): ...
-class PreprocessError(CCompilerError): ...
-class CompileError(CCompilerError): ...
-class LibError(CCompilerError): ...
-class LinkError(CCompilerError): ...
-class UnknownFileError(CCompilerError): ...
